+++ src/backend/joanie/core/utils.py
@@ -1,0 +1,9 @@
+"""
+Utils that can be useful throughout Joanie's core app
+"""
+from django.utils.text import slugify
+
+
+def normalize_code(code):
+    """Normalize object codes to avoid duplicates."""
+    return slugify(code, allow_unicode=False).upper() if code else None